ci: annotate errors: extract GitHub logic
